+++ src/locale/locale.js
@@ -70,6 +70,7 @@ export default class Locale {
 		this.ampersand = "&";
 		this.sectionSign = "§";
 		this.copyright = "©";
+		this.registeredTrademark = "®";
 		this.soundRecordingCopyright = "℗";
 		this.trademark = "™";