+++ src/typopo.js
@@ -22,6 +22,7 @@ import {fixMultiplicationSign} from "./lib/symbols/multiplication-sign";
 import {fixSectionSign} from "./lib/symbols/section-sign";
 import {fixCopyright} from "./lib/symbols/copyright";
 import {fixSoundRecordingCopyright} from "./lib/symbols/sound-recording-copyright";
+import {fixRegisteredTrademark} from "./lib/symbols/registered-trademark";
 import {fixTrademark} from "./lib/symbols/trademark";
 import {fixExponents} from "./lib/symbols/exponents";
 import {fixAbbreviations} from "./lib/words/abbreviations";
@@ -72,6 +73,7 @@ export function fixTypos(string, locale, configuration) {
 	string = fixSectionSign(string, currentLocale);
 	string = fixCopyright(string, currentLocale);
 	string = fixSoundRecordingCopyright(string, currentLocale);
+	string = fixRegisteredTrademark(string, currentLocale);
 	string = fixTrademark(string, currentLocale);
 	string = fixExponents(string, currentLocale);